+++ src/main/java/shadows/apotheosis/Apotheosis.java
@@ -27,10 +27,8 @@ import net.minecraftforge.network.NetworkRegistry;
 import net.minecraftforge.network.simple.SimpleChannel;
 import shadows.apotheosis.advancements.AdvancementTriggers;
 import shadows.apotheosis.compat.PatchouliCompat;
-import shadows.apotheosis.compat.TOPCompat;
 import shadows.apotheosis.ench.EnchModule;
 import shadows.apotheosis.ench.table.ClueMessage;
-import shadows.apotheosis.ench.table.EnchantingStatManager.StatSyncMessage;
 import shadows.apotheosis.garden.GardenModule;
 import shadows.apotheosis.potion.PotionModule;
 import shadows.apotheosis.spawn.SpawnerModule;
@@ -106,14 +104,12 @@ public class Apotheosis {
 		bus.addListener(this::init);
 		MinecraftForge.EVENT_BUS.addListener(this::trackCooldown);
 		MinecraftForge.EVENT_BUS.addListener(this::reloads);
-		if (ModList.get().isLoaded("theoneprobe")) TOPCompat.register();
 		if (ModList.get().isLoaded("patchouli")) PatchouliCompat.register();
 	}
 
 	@SubscribeEvent
 	public void init(FMLCommonSetupEvent e) {
 		MessageHelper.registerMessage(CHANNEL, 0, new ParticleMessage());
-		MessageHelper.registerMessage(CHANNEL, 1, new StatSyncMessage());
 		MessageHelper.registerMessage(CHANNEL, 2, new ClueMessage(0, null, false));
 		e.enqueueWork(AdvancementTriggers::init);
 		CraftingHelper.register(new ModuleCondition.Serializer());