+++ src/main/java/shadows/apotheosis/ench/compat/EnchTOPPlugin.java
@@ -8,8 +8,8 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
-import shadows.apotheosis.compat.TOPCompat;
 import shadows.apotheosis.ench.anvil.AnvilTile;
+import shadows.placebo.compat.TOPCompat;
 
 public class EnchTOPPlugin implements TOPCompat.Provider {