+++ src/main/java/shadows/apotheosis/village/wanderer/JsonTrade.java
@@ -1,0 +1,10 @@
+package shadows.apotheosis.village.wanderer;
+
+import net.minecraft.world.entity.npc.VillagerTrades.ItemListing;
+import shadows.placebo.json.PlaceboJsonReloadListener.TypeKeyed;
+
+public interface JsonTrade extends ItemListing, TypeKeyed<JsonTrade> {
+
+	boolean isRare();
+
+}