+++ src/main/java/shadows/apotheosis/village/wanderer/WandererTradeManager.java
@@ -5,68 +5,65 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonElement;
-
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.packs.resources.ResourceManager;
-import net.minecraft.server.packs.resources.SimpleJsonResourceReloadListener;
-import net.minecraft.util.profiling.ProfilerFiller;
+import net.minecraft.util.GsonHelper;
+import net.minecraft.world.entity.npc.VillagerTrades.ItemListing;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.trading.MerchantOffer;
 import net.minecraftforge.common.BasicItemListing;
-import shadows.apotheosis.util.JsonUtil;
+import shadows.apotheosis.Apotheosis;
 import shadows.apotheosis.village.VillageModule;
 import shadows.placebo.json.ItemAdapter;
-import shadows.placebo.json.NBTAdapter;
+import shadows.placebo.json.PlaceboJsonReloadListener;
+import shadows.placebo.json.SerializerBuilder;
 
-public class WandererTradeManager extends SimpleJsonResourceReloadListener {
-
-	public static final Gson GSON = new GsonBuilder().setPrettyPrinting().registerTypeAdapter(BasicItemListing.class, BasicItemListingAdapter.INSTANCE).registerTypeAdapter(ItemStack.class, ItemAdapter.INSTANCE).registerTypeAdapter(CompoundTag.class, NBTAdapter.INSTANCE).create();
+public class WandererTradeManager extends PlaceboJsonReloadListener<JsonTrade> {
 
 	public static final WandererTradeManager INSTANCE = new WandererTradeManager();
 
 	protected final Map<ResourceLocation, BasicItemListing> registry = new HashMap<>();
-	protected final List<BasicItemListing> normTrades = new ArrayList<>();
-	protected final List<BasicItemListing> rareTrades = new ArrayList<>();
+	protected final List<ItemListing> normTrades = new ArrayList<>();
+	protected final List<ItemListing> rareTrades = new ArrayList<>();
 
 	public WandererTradeManager() {
-		super(GSON, "wanderer_trades");
+		super(VillageModule.LOGGER, "wanderer_trades", false, true);
 	}
 
 	@Override
-	protected void apply(Map<ResourceLocation, JsonElement> objects, ResourceManager mgr, ProfilerFiller profiler) {
-		this.normTrades.clear();
-		this.rareTrades.clear();
-		this.registry.clear();
-		objects.forEach((id, obj) -> {
-			try {
-				if (!JsonUtil.checkAndLogEmpty(obj, id, "Wanderer Trade", VillageModule.LOGGER)) this.register(id, GSON.fromJson(obj, BasicItemListing.class), obj.getAsJsonObject().has("rare"));
-			} catch (Exception e) {
-				VillageModule.LOGGER.error("Failed to load Wandering Trader trade {}.", id.toString());
-				e.printStackTrace();
-			}
-		});
-		VillageModule.LOGGER.info("Loaded {} normal and {} rare Wandering Trader trade options.", this.normTrades.size(), this.rareTrades.size());
+	protected void registerBuiltinSerializers() {
+		this.registerSerializer(new ResourceLocation(Apotheosis.MODID, "basic_trade"), new SerializerBuilder<JsonTrade>("Basic JSON Trade").withJsonDeserializer(obj -> {
+			ItemStack price1 = ItemAdapter.ITEM_READER.fromJson(obj.get("input_1"), ItemStack.class);
+			ItemStack price2 = obj.has("input_2") ? ItemAdapter.ITEM_READER.fromJson(obj.get("input_2"), ItemStack.class) : ItemStack.EMPTY;
+			ItemStack output = ItemAdapter.ITEM_READER.fromJson(obj.get("output"), ItemStack.class);
+			int maxTrades = GsonHelper.getAsInt(obj, "max_trades", 1);
+			int xp = GsonHelper.getAsInt(obj, "xp", 0);
+			float priceMult = GsonHelper.getAsFloat(obj, "price_mult", 1);
+			boolean rare = GsonHelper.getAsBoolean(obj, "rare", false);
+			return new BasicJsonTrade(price1, price2, output, maxTrades, xp, priceMult, rare);
+		}));
 	}
 
-	protected void register(ResourceLocation id, BasicItemListing trade, boolean rare) {
+	@Override
+	protected <T extends JsonTrade> void register(ResourceLocation key, T trade) {
 		MerchantOffer offer = trade.getOffer(null, null);
-		if (offer.getResult() == null || offer.getResult().isEmpty() || offer.getMaxUses() == 0) return;
-		if (!this.registry.containsKey(id)) {
-			this.registry.put(id, trade);
-			if (rare) this.rareTrades.add(trade);
-			else this.normTrades.add(trade);
-		} else VillageModule.LOGGER.error("Attempted to register a wanderer trade with name {}, but it already exists!", id);
+		if (offer.getResult() == null || offer.getResult().isEmpty() || offer.getMaxUses() == 0) throw new RuntimeException("Wanderer Trade " + key + " is invalid.");
+		super.register(key, trade);
 	}
 
-	public List<BasicItemListing> getNormalTrades() {
+	@Override
+	protected void onReload() {
+		super.onReload();
+		this.getValues().forEach(trade -> {
+			if (trade.isRare()) this.rareTrades.add(trade);
+			else this.normTrades.add(trade);
+		});
+	}
+
+	public List<ItemListing> getNormalTrades() {
 		return this.normTrades;
 	}
 
-	public List<BasicItemListing> getRareTrades() {
+	public List<ItemListing> getRareTrades() {
 		return this.rareTrades;
 	}